Broke apart toolbar into separate components.
