+++ src/renderer/components/Toolbar.tsx
@@ -1,95 +1,5 @@
-// Icons
-import NextIcon from '@mui/icons-material/FastForward';
-import PlayArrowIcon from '@mui/icons-material/PlayArrow';
-import StopIcon from '@mui/icons-material/Stop';
-
-// Material UI.
-import { Box, FormControl, IconButton, InputLabel, MenuItem, Select, Stack, Tooltip, Typography, useTheme } from '@mui/material';
-import { Miner, minerState$ } from 'models';
-
-// Services.
-import { getAppSettings, setAppSettings, watchers$ } from 'renderer/services/AppSettingsService';
-import * as formatter from 'renderer/services/Formatters';
-import { nextCoin, startMiner, stopMiner } from 'renderer/services/MinerManager';
-import { minerStatistics$ } from 'renderer/services/StatisticsAggregator';
-
-// Hooks.
-import { useProfile, useObservableState, useLoadData } from '../hooks';
-
-function Separator() {
-  const theme = useTheme();
-  return <span style={{ fontSize: '1.4rem', fontWeight: 'lighter', color: theme.palette.text.disabled, margin: '0 0.2rem' }}>|</span>;
-}
-
-function MinerSummary() {
-  const [minerState] = useObservableState(minerState$, null);
-  const [minerStatistic] = useObservableState(minerStatistics$, null);
-  const minerActive = minerState?.state === 'active';
-
-  return (
-    <Box>
-      {minerActive && (
-        <Typography sx={{ mr: 2 }}>
-          <strong>Coin</strong>: {minerState?.currentCoin} <Separator /> <strong>Hashrate</strong>: {formatter.hashrate(minerStatistic?.hashrate)}{' '}
-        </Typography>
-      )}
-    </Box>
-  );
-}
-
-function MinerSelector() {
-  const [miners, setLoadedMiners] = useObservableState(watchers$.miners, []);
-
-  useLoadData(async ({ getMiners }) => {
-    getMiners()
-      .then(setLoadedMiners)
-      .catch((err) => console.error('Failed to load miners: ', err));
-  });
-
-  const profile = useProfile();
-
-  const setDefaultMiner = async (name: string) => {
-    const appSettings = await getAppSettings();
-    await setAppSettings({ ...appSettings, settings: { ...appSettings.settings, defaultMiner: name } });
-  };
-
-  return (
-    <FormControl size="small" sx={{ minWidth: '12rem' }}>
-      <InputLabel id="miner-label">Miner</InputLabel>
-      <Select labelId="miner-label" sx={{ fontSize: '0.8rem' }} label="Miner" value={profile ?? ''} onChange={($event) => setDefaultMiner($event.target.value)}>
-        {(miners ?? Array<Miner>())
-          .sort((a, b) => a.name.localeCompare(b.name))
-          .map((miner) => (
-            <MenuItem key={miner.name} value={miner.name}>
-              {miner.name} ({miner.kind})
-            </MenuItem>
-          ))}
-      </Select>
-    </FormControl>
-  );
-}
-
-function MinerControls() {
-  const profile = useProfile();
-  const [minerState] = useObservableState(minerState$, null);
-  const minerActive = minerState?.state === 'active';
-
-  return (
-    <Stack justifyContent="space-around" alignItems="center" direction="row">
-      <Tooltip title={minerActive ? 'Stop Miner' : 'Start Miner'}>
-        <IconButton onClick={() => (minerActive ? stopMiner() : startMiner())}>{minerActive ? <StopIcon color="error" /> : <PlayArrowIcon color="primary" />}</IconButton>
-      </Tooltip>
-      <Separator />
-      <Tooltip title="Next Coin">
-        <span>
-          <IconButton disabled={!minerActive || !profile} onClick={() => nextCoin()}>
-            <NextIcon />
-          </IconButton>
-        </span>
-      </Tooltip>
-    </Stack>
-  );
-}
+import { Box, Stack, useTheme } from '@mui/material';
+import { MinerSummary, MinerSelector, MinerControls } from './toolbars';
 
 export function Toolbar({ drawerWidth }: { drawerWidth: number }) {
   const theme = useTheme();
