+++ src/renderer/components/toolbars/index.ts
@@ -1,0 +1,3 @@
+export { MinerSummary } from './MinerSummary';
+export { MinerSelector } from './MinerSelector';
+export { MinerControls } from './MinerControls';
